requirements.txt
tavily-python>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
